+++ .devcontainer/hooks/onCreate.sh
@@ -2,7 +2,8 @@
 set -euo pipefail
 
 WORKSPACE_DIR=$(realpath /workspaces/*)
+INTERNAL_DIR="$WORKSPACE_DIR/.internal"
 
-npm ci
+npm ci --prefix "$INTERNAL_DIR"
 npm ci --prefix "$HOME/twinejs"
-poetry install -P "$WORKSPACE_DIR/spindler"
+poetry install -P "$INTERNAL_DIR/spindler"